Updated ini check for loose file within setup
